Remove expireDeadHosts in BlockManagerMasterActor
